add processTemplateString
